+++ tik_manager/SmPhotoshop.py
@@ -88,132 +88,12 @@ class PsManager(RootManager):
         # self.psApp = ct.CreateObject('Photoshop.Application')
         self.psApp = Dispatch('Photoshop.Application')
 
-    # def init_paths(self):
-    #     """Overriden function"""
-    #     logger.debug("Func: init_paths")
-    #     _softwarePathsDict = self.getSoftwarePaths()
-    #     self._pathsDict["userSettingsDir"] = os.path.normpath(os.path.join(self.getUserDirectory(), _softwarePathsDict["userSettingsDir"]))
-    #     self._folderCheck(self._pathsDict["userSettingsDir"])
-    #
-    #     self._pathsDict["bookmarksFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smBookmarks.json"))
-    #     self._pathsDict["currentsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smCurrents.json"))
-    #     self._pathsDict["projectsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smProjects.json"))
-    #
-    #     self._pathsDict["commonFolder"] = os.path.abspath(os.path.join(self._pathsDict["userSettingsDir"], os.pardir))
-    #     self._pathsDict["commonFolderFile"] = os.path.normpath(os.path.join(self._pathsDict["commonFolder"], "smCommonFolder.json"))
-    #
-    #     self._pathsDict["projectDir"] = self.getProjectDir()
-    #     self._pathsDict["sceneFile"] = ""
-    #
-    #     self._pathsDict["masterDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], "smDatabase"))
-    #     self._folderCheck(self._pathsDict["masterDir"])
-    #
-    #     self._pathsDict["databaseDir"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], _softwarePathsDict["databaseDir"]))
-    #     self._folderCheck(self._pathsDict["databaseDir"])
-    #
-    #     self._pathsDict["scenesDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], _softwarePathsDict["scenesDir"]))
-    #     self._folderCheck(self._pathsDict["scenesDir"])
-    #
-    #     self._pathsDict["projectSettingsFile"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], "projectSettings.json"))
-    #
-    #     self._pathsDict["subprojectsFile"] = os.path.normpath(os.path.join(self._pathsDict["masterDir"], "subPdata.json"))
-    #     self._pathsDict["categoriesFile"] = os.path.normpath(os.path.join(self._pathsDict["databaseDir"], _softwarePathsDict["categoriesFile"]))
-    #     self._pathsDict["previewsDir"] = os.path.normpath(os.path.join(self._pathsDict["projectDir"], "Playblasts", _softwarePathsDict["niceName"])) # dont change
-    #     self._pathsDict["pbSettingsFile"] = os.path.normpath(os.path.join(self._pathsDict["previewsDir"], _softwarePathsDict["pbSettingsFile"]))
-    #
-    #     self._pathsDict["generalSettingsDir"] = self._getCommonFolder()
-    #     if self._pathsDict["generalSettingsDir"] == -1:
-    #         self._exception(201, "Cannot Continue Without Common Database")
-    #         return -1
-    #
-    #     self._pathsDict["usersFile"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "sceneManagerUsers.json"))
-    #
-    #     self._pathsDict["softwareDatabase"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "softwareDatabase.json"))
-    #     self._pathsDict["sceneManagerDefaults"] = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "sceneManagerDefaults.json"))
-
-    def _checkCommonFolder(self, folder):
-        checkList = [os.path.join(folder, "sceneManagerDefaults.json"),
-                     os.path.join(folder, "sceneManagerUsers.json"),
-                     os.path.join(folder, "softwareDatabase.json")]
-        missingList = [os.path.basename(path) for path in checkList if not os.path.isfile(path)]
-        if len(missingList) > 0:
-
-            errorBox = QtWidgets.QMessageBox()
-            errorBox.setText("Common Database Folder missing some necessary files")
-            errorBox.setDetailedText("Following files are missing:\n %s" %pprint.pformat(missingList))
-            errorBox.exec_()
-            return False
-        else:
-            return True
-
-    def _getCommonFolder(self):
-        """Standalone Specific function - Different from software modules,
-        Standalone version needs to know where common files exist"""
-        if os.path.isfile(self._pathsDict["commonFolderFile"]):
-            commonFolder = self._loadJson(self._pathsDict["commonFolderFile"])
-            if commonFolder == -2:
-                return -2
-
-        else:
-            # in case there is no file found (Initial Run)
-            q = QtWidgets.QMessageBox()
-            q.setIcon(QtWidgets.QMessageBox.Question)
-            q.setText("Please define the Common Database Folder which contains general database files\n\nPress Ok to continue")
-            q.setWindowTitle("Select the Common Database Folder")
-            q.setDetailedText(
-                "Common Database Folder is the one which has the common modules for all software modules. Common Database Folder must include:\n"
-                "sceneManagerDefaults.json\n"
-                "sceneManagerUsers.json\n"
-                "softwareDatabase.json")
-            q.setStandardButtons(
-                QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Abort)
-
-            q.button(QtWidgets.QMessageBox.Ok).setFixedWidth(100)
-            #
-            q.button(QtWidgets.QMessageBox.Abort).setFixedWidth(100)
-            ret = q.exec_()
-            if ret == QtWidgets.QMessageBox.Ok:
-                commonFolder = self._defineCommonFolder()
-
-            elif ret == QtWidgets.QMessageBox.Abort:
-                return -1
-
-        return commonFolder
-
-    def _defineCommonFolder(self):
-        dlg = QtWidgets.QFileDialog.getExistingDirectory()
-        if dlg:
-            selectedDir = os.path.normpath(str(dlg))
-            if self._checkCommonFolder(selectedDir):
-                commonFolder = selectedDir
-                self._saveCommonFolder(commonFolder)
-
-                q = QtWidgets.QMessageBox()
-                q.setIcon(QtWidgets.QMessageBox.Information)
-                q.setText("Common Database Defined Successfully")
-                q.setWindowTitle("Success")
-                q.exec_()
-
-                return commonFolder
-            else:
-                return self._getCommonFolder()
-        else:
-            return self._getCommonFolder()
-
-    def _saveCommonFolder(self, data):
-        try:
-            self._dumpJson(data, self._pathsDict["commonFolderFile"])
-            msg = ""
-            return 0, msg
-        except:
-            msg = "Cannot save common folder file"
-            return -1, msg
-
     def getSoftwarePaths(self):
         """Overriden function"""
         logger.debug("Func: getSoftwarePaths")
 
         softwareDatabaseFile = os.path.normpath(os.path.join(os.path.dirname(os.path.abspath(__file__)), "softwareDatabase.json"))
+        # softwareDatabaseFile = os.path.normpath(os.path.join(self._pathsDict["generalSettingsDir"], "softwareDatabase.json"))
         softwareDB = self._loadJson(softwareDatabaseFile)
         # To tell the base class maya specific path names
         # print softwareDB
@@ -564,7 +444,10 @@ class PsManager(RootManager):
         filePath = os.path.join(exportFolderPath, fileName)
 
         if os.path.isfile(filePath):
-            if not self._query(("The following file will be overwritten if you continue:\n %s" %fileName), "Choose Yes to overwrite file and continue"):
+            header = "The following file will be overwritten if you continue:\n %s" %fileName
+            msg = "Choose Yes to overwrite file and continue"
+            title = "Are you Sure?"
+            if not self._question(header=header, msg=msg, title=title):
                 return
             else:
                 pass
@@ -708,7 +591,7 @@ class PsManager(RootManager):
                 new_height = oWidth / ratio
             dupDocument.ResizeImage(new_width, new_height)
             dupDocument.ResizeCanvas(oWidth, oHeight)
-            jpgSaveOptions = ct.CreateObject("Photoshop.JPEGSaveOptions")
+            jpgSaveOptions = Dispatch("Photoshop.JPEGSaveOptions")
             jpgSaveOptions.EmbedColorProfile = True
             jpgSaveOptions.FormatOptions = 1  # => psStandardBaseline
             jpgSaveOptions.Matte = 1  # => No Matte
@@ -760,24 +643,7 @@ class PsManager(RootManager):
             self._dumpJson(categoriesData, self._pathsDict["categoriesFile"])
         return categoriesData
 
-    def _query(self, header, msg):
-        q = QtWidgets.QMessageBox()
-        q.setIcon(QtWidgets.QMessageBox.Question)
-        q.setText(header)
-        q.setInformativeText(msg)
-        q.setWindowTitle("Are you sure?")
-        q.setStandardButtons(
-            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
-        q.button(QtWidgets.QMessageBox.Yes).setFixedHeight(30)
-        q.button(QtWidgets.QMessageBox.Yes).setFixedWidth(100)
-        q.button(QtWidgets.QMessageBox.Cancel).setFixedHeight(30)
-        q.button(QtWidgets.QMessageBox.Cancel).setFixedWidth(100)
 
-        ret = q.exec_()
-        if ret == QtWidgets.QMessageBox.Yes:
-            return True
-        else:
-            return False
 
 
 
@@ -792,6 +658,48 @@ class PsManager(RootManager):
 
         if (200 >= code < 210):
             raise Exception(code, msg)
+
+    def _question(self, header="", msg="", title="Manager Question"):
+        """OVERRIDEN METHOD"""
+        q = QtWidgets.QMessageBox()
+        q.setIcon(QtWidgets.QMessageBox.Question)
+        q.setText(header)
+        q.setInformativeText(msg)
+        q.setWindowTitle(title)
+        q.setStandardButtons(
+            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedWidth(100)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedWidth(100)
+
+        ret = q.exec_()
+        if ret == QtWidgets.QMessageBox.Yes:
+            return True
+        else:
+            return False
+
+    # def _question(self, msg):
+    #     """OVERRIDEN METHOD"""
+    #     state = cmds.confirmDialog( title='Manager Question', message=msg, button=['Yes','No'], defaultButton='Yes', cancelButton='No', dismissString='No' )
+    #     if state == "Yes":
+    #         return True
+    #     else:
+    #         return False
+
+    def _info(self, msg):
+        """OVERRIDEN METHOD"""
+        infobox = QtWidgets.QMessageBox()
+        infobox.setModal(True)
+        infobox.setText(msg)
+        infobox.setWindowTitle("Info")
+        infobox.exec_()
+
+    def _inputDir(self):
+        """OVERRIDEN METHOD"""
+        # Qt File dialog is preferred because it is faster
+        inputDir = str(QtWidgets.QFileDialog.getExistingDirectory())
+        return os.path.normpath(inputDir)
 
 
 class MainUI(baseUI):