+++ tik_manager/SmStandalone.py
@@ -64,62 +64,6 @@ logging.basicConfig()
 logger = logging.getLogger('smStandalone')
 logger.setLevel(logging.WARNING)
 
-# hard coded software dictionary for getting executables and creating viewer objects
-# softwareDictionary = {
-#     "Maya":{
-#         "niceName": "Maya",
-#         "root": "Autodesk",
-#         "relPath": "bin",
-#         "exeList": ["maya.exe"],
-#         "searchWord": "Maya",
-#         "databaseDir": "mayaDB",
-#         "scenesDir": "scenes",
-#         "pbSettingsFile": "pbSettings.json",
-#         "categoriesFile": "categoriesMaya.json",
-#         "userSettingsDir": "SceneManager\\Maya"
-#     },
-#     "3dsMax":
-#         {
-#             "niceName": "3dsMax",
-#             "root": "Autodesk",
-#             "relPath": "",
-#             "exeList": ["3dsmax.exe"],
-#             "searchWord": "3ds",
-#             "databaseDir": "maxDB",
-#             "scenesDir": "scenes_3dsMax",
-#             "pbSettingsFile": "pbSettings_3dsMax.json",
-#             "categoriesFile": "categories3dsMax.json",
-#             "userSettingsDir": "SceneManager\\3dsMax"
-#         },
-#     "Houdini":
-#         {
-#             "niceName": "Houdini",
-#             "root": "Side Effects Software",
-#             "relPath": "bin",
-#             "exeList": ["houdini.exe", "houdinifx.exe"],
-#             "searchWord": "Hou",
-#             "databaseDir": "houdiniDB",
-#             "scenesDir": "scenes_houdini",
-#             "pbSettingsFile": "pbSettings_houdini.json",
-#             "categoriesFile": "categoriesHoudini.json",
-#             "userSettingsDir": "SceneManager\\Houdini"
-#         },
-#     "Nuke":
-#         {
-#             "niceName": "Nuke",
-#             "root": "",
-#             "relPath": "",
-#             "exeList": ["Nuke10.0.exe","Nuke10.1.exe","Nuke10.2.exe","Nuke10.3.exe","Nuke10.4.exe","Nuke10.5.exe","Nuke10.6.exe","Nuke10.7.exe","Nuke10.8.exe","Nuke10.9.exe","Nuke11.0.exe","Nuke11.1.exe","Nuke11.2.exe","Nuke11.3.exe","Nuke11.4.exe","Nuke11.5.exe","Nuke11.6.exe","Nuke11.7.exe","Nuke11.8.exe","Nuke11.9.exe",],
-#             "searchWord": "Nuke",
-#             "databaseDir": "nukeDB",
-#             "scenesDir": "scenes_nuke",
-#             "pbSettingsFile": "pbSettings_nuke.json",
-#             "categoriesFile": "categoriesNuke.json",
-#             "userSettingsDir": "SceneManager\\Nuke"
-#         }
-# }
-
-
 class SwViewer(RootManager):
     """
     Helper class for StandaloneManager. Views and opens Scenes.
@@ -444,6 +388,22 @@ class SwViewer(RootManager):
             self._exception(360, msg)
             return
 
+        elif swID == "Photoshop":
+            # use the last available version for photoshop
+            if len(executables["Photoshop"]["64Bit"]) != 0:
+                exePath = executables["Photoshop"]["64Bit"][-1][1] # this is the last one in the list
+                subprocess.Popen([exePath, self.currentScenePath], shell=True)
+                return
+
+            if len(executables["Photoshop"]["32Bit"]) != 0:
+                exePath = executables["Photoshop"]["32Bit"][-1][1] # this is the last one in the list
+                subprocess.Popen([exePath, self.currentScenePath], shell=True)
+                return
+
+            msg = "Photoshop is not installed on this workstation"
+            self._exception(360, msg)
+            return
+
     def _exception(self, code, msg):
         """OVERRIDEN"""
 
@@ -461,42 +421,6 @@ class StandaloneManager(RootManager):
     """Command Class. Inherits SmRoot.py"""
     def __init__(self):
         super(StandaloneManager, self).__init__()
-
-        # Dictionary with items for each software supported by Scene Manager
-        # self.swDictList = [{"niceName": "3dsMax",
-        #                     "databaseDir": "maxDB",
-        #                     "scenesDir": "scenes_3dsMax",
-        #                     "pbSettingsFile": "pbSettings_3dsMax.json",
-        #                     "categoriesFile": "categories3dsMax.json",
-        #                     "userSettingsDir": "Documents\\SceneManager\\3dsMax"},
-        #
-        #                    {"niceName": "Maya",
-        #                     "databaseDir": "mayaDB",
-        #                     "scenesDir": "scenes",
-        #                     "pbSettingsFile": "pbSettings.json",
-        #                     "categoriesFile": "categoriesMaya.json",
-        #                     "userSettingsDir": "Documents\\SceneManager\\Maya"},
-        #
-        #                    {"niceName": "Houdini",
-        #                     "databaseDir": "houdiniDB",
-        #                     "scenesDir": "scenes_houdini",
-        #                     "pbSettingsFile": "pbSettings_houdini.json",
-        #                     "categoriesFile": "categoriesHoudini.json",
-        #                     "userSettingsDir": "Documents\\SceneManager\\Houdini"}
-        #                    ]
-
-        # self._pathsDict["userSettingsDir"] = os.path.normpath(os.path.join(os.path.expanduser("~"), "Documents", "SceneManager", "Standalone"))
-        # self._folderCheck(self._pathsDict["userSettingsDir"])
-        #
-        # self._pathsDict["bookmarksFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smBookmarks.json"))
-        # self._pathsDict["currentsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smCurrents.json"))
-        # self._pathsDict["projectsFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smProjects.json"))
-        #
-        # self._pathsDict["commonFolderFile"] = os.path.normpath(os.path.join(self._pathsDict["userSettingsDir"], "smCommonFolder.json"))
-        # self._pathsDict["generalSettingsDir"] = self._getCommonFolder()
-        # if self._pathsDict["generalSettingsDir"] == -1:
-        #     self._exception(201, "Cannot Continue Without Common Database")
-        #     return
 
         self.swList = []
         self.init_paths()
@@ -654,88 +578,122 @@ class StandaloneManager(RootManager):
         if (200 >= code < 210):
             raise Exception(code, msg)
 
-    def _checkCommonFolder(self, folder):
-        checkList = [os.path.join(folder, "sceneManagerDefaults.json"),
-                     os.path.join(folder, "sceneManagerUsers.json"),
-                     os.path.join(folder, "softwareDatabase.json")]
-        missingList = [os.path.basename(path) for path in checkList if not os.path.isfile(path)]
-        if len(missingList) > 0:
+    def _question(self, header="", msg="", title="Manager Question"):
+        """OVERRIDEN METHOD"""
+        q = QtWidgets.QMessageBox()
+        q.setIcon(QtWidgets.QMessageBox.Question)
+        q.setText(header)
+        q.setInformativeText(msg)
+        q.setWindowTitle(title)
+        q.setStandardButtons(
+            QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Yes).setFixedWidth(100)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedHeight(30)
+        q.button(QtWidgets.QMessageBox.Cancel).setFixedWidth(100)
 
-            errorBox = QtWidgets.QMessageBox()
-            errorBox.setText("Common Database Folder missing some necessary files")
-            errorBox.setDetailedText("Following files are missing:\n %s" %pprint.pformat(missingList))
-            errorBox.exec_()
-            return False
-        else:
+        ret = q.exec_()
+        if ret == QtWidgets.QMessageBox.Yes:
             return True
-
-    def _getCommonFolder(self):
-        """Standalone Specific function - Different from software modules,
-        Standalone version needs to know where common files exist"""
-        if os.path.isfile(self._pathsDict["commonFolderFile"]):
-            commonFolder = self._loadJson(self._pathsDict["commonFolderFile"])
-            if commonFolder == -2:
-                return -2
-
         else:
-            # in case there is no file found (Initial Run)
-            q = QtWidgets.QMessageBox()
-            q.setIcon(QtWidgets.QMessageBox.Question)
-            q.setText("Please define the Common Database Folder which contains general database files\n\nPress Ok to continue")
-            # q.setInformativeText(textInfo)
-            q.setWindowTitle("Select the Common Database Folder")
-            q.setDetailedText(
-                "Common Database Folder is the one which has the common modules for all software modules. Common Database Folder must include:\n"
-                "sceneManagerDefaults.json\n"
-                "sceneManagerUsers.json\n"
-                "softwareDatabase.json")
-            q.setStandardButtons(
-                QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Abort)
+            return False
 
-            # q.button(QtWidgets.QMessageBox.Ok).setFixedHeight(30)
-            q.button(QtWidgets.QMessageBox.Ok).setFixedWidth(100)
-            #
-            # q.button(QtWidgets.QMessageBox.Abort).setFixedHeight(30)
-            q.button(QtWidgets.QMessageBox.Abort).setFixedWidth(100)
-            ret = q.exec_()
-            if ret == QtWidgets.QMessageBox.Ok:
-                commonFolder = self._defineCommonFolder()
+    def _info(self, msg):
+        """OVERRIDEN METHOD"""
+        infobox = QtWidgets.QMessageBox()
+        infobox.setModal(True)
+        infobox.setText(msg)
+        infobox.setWindowTitle("Info")
+        infobox.exec_()
 
-            elif ret == QtWidgets.QMessageBox.Abort:
-                return -1
+    def _inputDir(self):
+        """OVERRIDEN METHOD"""
+        # Qt File dialog is preferred because it is faster
+        inputDir = str(QtWidgets.QFileDialog.getExistingDirectory())
+        return os.path.normpath(inputDir)
 
-        return commonFolder
-
-    def _defineCommonFolder(self):
-        dlg = QtWidgets.QFileDialog.getExistingDirectory()
-        if dlg:
-            selectedDir = os.path.normpath(str(dlg))
-            if self._checkCommonFolder(selectedDir):
-                commonFolder = selectedDir
-                self._saveCommonFolder(commonFolder)
-
-                q = QtWidgets.QMessageBox()
-                q.setIcon(QtWidgets.QMessageBox.Information)
-                q.setText("Common Database Defined Successfully")
-                # q.setInformativeText(textInfo)
-                q.setWindowTitle("Success")
-                q.exec_()
-
-                return commonFolder
-            else:
-                return self._getCommonFolder()
-        else:
-            return self._getCommonFolder()
-            # self._getCommonFolder()
-
-    def _saveCommonFolder(self, data):
-        try:
-            self._dumpJson(data, self._pathsDict["commonFolderFile"])
-            msg = ""
-            return 0, msg
-        except:
-            msg = "Cannot save common folder file"
-            return -1, msg
+    # def _checkCommonFolder(self, folder):
+    #     checkList = [os.path.join(folder, "sceneManagerDefaults.json"),
+    #                  os.path.join(folder, "sceneManagerUsers.json"),
+    #                  os.path.join(folder, "softwareDatabase.json")]
+    #     missingList = [os.path.basename(path) for path in checkList if not os.path.isfile(path)]
+    #     if len(missingList) > 0:
+    #
+    #         errorBox = QtWidgets.QMessageBox()
+    #         errorBox.setText("Common Database Folder missing some necessary files")
+    #         errorBox.setDetailedText("Following files are missing:\n %s" %pprint.pformat(missingList))
+    #         errorBox.exec_()
+    #         return False
+    #     else:
+    #         return True
+    #
+    # def _getCommonFolder(self):
+    #     """Standalone Specific function - Different from software modules,
+    #     Standalone version needs to know where common files exist"""
+    #     if os.path.isfile(self._pathsDict["commonFolderFile"]):
+    #         commonFolder = self._loadJson(self._pathsDict["commonFolderFile"])
+    #         if commonFolder == -2:
+    #             return -2
+    #
+    #     else:
+    #         # in case there is no file found (Initial Run)
+    #         q = QtWidgets.QMessageBox()
+    #         q.setIcon(QtWidgets.QMessageBox.Question)
+    #         q.setText("Please define the Common Database Folder which contains general database files\n\nPress Ok to continue")
+    #         # q.setInformativeText(textInfo)
+    #         q.setWindowTitle("Select the Common Database Folder")
+    #         q.setDetailedText(
+    #             "Common Database Folder is the one which has the common modules for all software modules. Common Database Folder must include:\n"
+    #             "sceneManagerDefaults.json\n"
+    #             "sceneManagerUsers.json\n"
+    #             "softwareDatabase.json")
+    #         q.setStandardButtons(
+    #             QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Abort)
+    #
+    #         # q.button(QtWidgets.QMessageBox.Ok).setFixedHeight(30)
+    #         q.button(QtWidgets.QMessageBox.Ok).setFixedWidth(100)
+    #         #
+    #         # q.button(QtWidgets.QMessageBox.Abort).setFixedHeight(30)
+    #         q.button(QtWidgets.QMessageBox.Abort).setFixedWidth(100)
+    #         ret = q.exec_()
+    #         if ret == QtWidgets.QMessageBox.Ok:
+    #             commonFolder = self._defineCommonFolder()
+    #
+    #         elif ret == QtWidgets.QMessageBox.Abort:
+    #             return -1
+    #
+    #     return commonFolder
+    #
+    # def _defineCommonFolder(self):
+    #     dlg = QtWidgets.QFileDialog.getExistingDirectory()
+    #     if dlg:
+    #         selectedDir = os.path.normpath(str(dlg))
+    #         if self._checkCommonFolder(selectedDir):
+    #             commonFolder = selectedDir
+    #             self._saveCommonFolder(commonFolder)
+    #
+    #             q = QtWidgets.QMessageBox()
+    #             q.setIcon(QtWidgets.QMessageBox.Information)
+    #             q.setText("Common Database Defined Successfully")
+    #             # q.setInformativeText(textInfo)
+    #             q.setWindowTitle("Success")
+    #             q.exec_()
+    #
+    #             return commonFolder
+    #         else:
+    #             return self._getCommonFolder()
+    #     else:
+    #         return self._getCommonFolder()
+    #         # self._getCommonFolder()
+    #
+    # def _saveCommonFolder(self, data):
+    #     try:
+    #         self._dumpJson(data, self._pathsDict["commonFolderFile"])
+    #         msg = ""
+    #         return 0, msg
+    #     except:
+    #         msg = "Cannot save common folder file"
+    #         return -1, msg
 
     def _loadUserPrefs(self):
         """OVERRIDEN FUNCTION Load Last CategoryIndex, SubProject Index,